requirements.txt
Flask==1.1.1
line-bot-sdk==1.15.0
requests==2.23.0
beautifulsoup4==4.8.2
urllib3==1.25.8
gunicorn==20.0.4
google-cloud-storage==1.19.0
Pillow==6.1.0
pytorch==1.4.0
opencv-python
datetime